# api-rest-amazonia/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/municipios": {
      "get": {
        "operationId": "MunicipiosController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de la página (empieza en 1). Si no se envía, se toma por defecto el valor 1.",
            "schema": {
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Cantidad de resultados por página. Si no se envía, se toma por defecto el valor 10.",
            "schema": {
              "maximum": 100,
              "default": 10,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "departamento",
            "required": false,
            "in": "query",
            "description": "Filtrar por ID de departamento",
            "schema": {
              "example": 3,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Buscar por nombre de municipio (búsqueda parcial, insensible a mayúsculas)",
            "schema": {
              "example": "Trinidad",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Listado paginado de municipios",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponseDto"
                }
              }
            }
          }
        },
        "summary": "Listar municipios con paginación y filtros opcionales",
        "tags": [
          "Municipios"
        ]
      }
    },
    "/api/municipios/{id}/comunidades": {
      "get": {
        "operationId": "MunicipiosController_findComunidades",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del municipio",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de comunidades indígenas del municipio, ordenadas por nombre"
          }
        },
        "summary": "Comunidades indígenas de un municipio (carga en cascada para formularios)",
        "tags": [
          "Municipios"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kaa Iya — Backend API",
    "description": "API REST de la plataforma **Kaa Iya** (\"Espíritu del Bosque\"), para gestionar y visibilizar iniciativas sostenibles en la Amazonía boliviana.\n\n**Base URL:** `/api`\n\n**Autenticación:** Bearer JWT. Obtén tu token en `POST /auth/login` e inclúyelo en todas las peticiones protegidas: `Authorization: Bearer <token>`",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo Kaa Iya — UCB San Pablo",
      "url": "https://www.ucb.edu.bo",
      "email": "catedra.amazonia@ucb.edu.bo"
    },
    "license": {
      "name": "UNLICENSED",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development"
    },
    {
      "url": "https://kaaiya-backend.onrender.com",
      "description": "Producción (Render/Railway — ajustar al dominio real)"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Ingresa el token JWT obtenido de POST /auth/login"
      }
    },
    "schemas": {
...
      "PaginationResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Array de elementos de la página actual",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "page": {
            "type": "number",
            "description": "Número de la página actual",
            "example": 1
          },
          "limit": {
            "type": "number",
            "description": "Cantidad de elementos por página",
            "example": 10
          },
          "pages": {
            "type": "number",
            "description": "Cantidad total de páginas disponibles",
            "example": 5
          },
          "total": {
            "type": "number",
            "description": "Cantidad total de registros encontrados",
            "example": 42
          },
          "has_next": {
            "type": "boolean",
            "description": "Indica si existe una página siguiente",
            "example": true
          },
          "has_prev": {
            "type": "boolean",
            "description": "Indica si existe una página anterior",
            "example": false
          }
        },
        "required": [
          "data",
          "page",
          "limit",
          "pages",
          "total",
          "has_next",
          "has_prev"
        ]
      },
...
    }
  }
}